Added dialects
